% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_stars.R
\name{add_stars}
\alias{add_stars}
\title{Add Significance Stars}
\usage{
add_stars(pval, alpha = c(0.001, 0.01, 0.05))
}
\arguments{
\item{pval}{A vector of pvalues}

\item{alpha}{a vector of three critical significance thresholds, defaults to
\code{c(0.001, 0.01, 0.05)}.}
}
\description{
This function creates a character vector of \code{length(pval)}
adding significance stars according to the \code{alpha} values
specified.
}
\examples{
p <- c(0.1, 0.3, 0.0005, 0.02)
add_stars(p)
}
\keyword{hypothesis}
\keyword{p-values}
\keyword{significance,}
\keyword{testing,}
