% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_chisq.R
\name{batch_chisq}
\alias{batch_chisq}
\title{Batch Chi-square Tests}
\usage{
batch_chisq(df, DV, IV, ...)
}
\arguments{
\item{df}{The data frame that contains the \code{IV} and \code{DV}}

\item{DV}{A column name containing a factor variable, will be the columns
of the chi-sq contingecy tables.}

\item{IV}{A vector of column names of \code{length(IV) >= 1} contaiing
factor variables.}

\item{...}{pases options to the child functions, e.g. the number of
replicates to use in simulation \code{B = 1000}.}
}
\description{
A wrapper for \code{\link{chisq_wrap}} to enable batch processing
of Chi-square tests of association between one DV and a vector of
at least 1 IV.
}
\examples{
batch_chisq(df = enve_test, DV = "extortion_victim",
            IV = c("bribe_victim", "size"))
}
\keyword{chi-squared}
