% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_count.R
\name{chisq_count}
\alias{chisq_count}
\title{Chi-sq test of a count distribution}
\usage{
chisq_count(x, data = NULL, B = 500, family = c("poisson", "nbinom"))
}
\arguments{
\item{x}{a vector of count data or a name of a count data column in data}

\item{data}{a data frame}

\item{B}{nuber of replicates for simulating the distributions and
for calculating the p-value of the chi-squared test if any expected
counts are fewer than 5.}

\item{family}{families for the null hypothesis}
}
\description{
Performs a chi-squared test of a count distribution against
a vector of probabilities calculated via a monte carlo
simulation of distributions under the null hypothesis
of poisson or negative binomial expected counts.
}
\examples{
chisq_count(testdata$extortions, family = "poisson")

# batch using \\code{\\link{chisq_tb}}
lx <- lapply(c("extortions", "bribes"), chisq_count, testdata, B = 2000, family = "poisson")
chisq_tb(lx, stars = TRUE)
}
