% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_list.R
\name{chisq_list}
\alias{chisq_list}
\title{List of contingency tables}
\usage{
chisq_list(x, option = c("observed", "expected", "ratio", "percent"),
  print_option = c("none", "markdown", "pandoc", "latex", "html"))
}
\arguments{
\item{x}{A list of chi-squared tests.}

\item{option}{Allows specifying the type of continency table
to be produced. The options are:
- \code{observed} counts, default option.
- \code{expected} counts.
- \code{ratio} of observed to expected counts.
- \code{percent} row percentages.}

\item{print_option}{Allows specifying the output format, using
the \code{\link{kable}} from package \code{\link{knitr}}. The
default option \code{none} produces a list of matrix objects.}
}
\description{
A convenience function to output a list of contingency tables
from a list of chi-square tests, as those created by
\code{\link{batch_chisq}}.
}
\examples{
b <- batch_chisq(df = testdata, DV = "extortion_victim",
            IV = c("bribe_victim", "size"))

chisq_list(b, option = "ratio", print_option = "pandoc")
}
\keyword{chi-squared,}
\keyword{contingency}
\keyword{table}
