% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_tb.R
\name{chisq_tb}
\alias{chisq_tb}
\title{Table of chi-square tests results}
\usage{
chisq_tb(x, stars = TRUE, ...)
}
\arguments{
\item{x}{A list of \code{chisq.test} objects, as the one generated
by \code{batch_chisq}.}

\item{stars}{Option to add a column of stars indicating significance
using the \code{\link{add_stars}} function. Can pass a different
alpha vector of three thresholds using \code{...}.}
}
\description{
Generates a table summarising the results from a list of
\code{chisq.test} objects.
}
\examples{
b <- batch_chisq(enve_test, DV = "extortion_victim",
                  IV = c("bribe_victim", "size"))
chisq_tb(b)
}
\keyword{chi-squared}
\keyword{table}
\keyword{test,}
