% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_hurdle2other.R
\name{compare_hurdle2other}
\alias{compare_hurdle2other}
\title{compare a glmmadmb hurdle to non-hurdle models}
\usage{
compare_hurdle2other(mlogit, mcount, ..., modnames = c("Hurdle", "Negbin.",
  "ZINB"))
}
\arguments{
\item{mlogit}{a binary model}

\item{mcount}{a truncated count model}

\item{...}{a series of at least one comparable glmmadmb model}

\item{modnames}{a vector of names  for the models supplied. \code{mlogit}
and \code{mcount} correspond to "Hurdle", while the rest of the models
supplied in \code{...} correspond to the rest of the names. If only one
model is supplied in \code{...}, only the first two names are used.}
}
\description{
This function compares the joint parameters of a hurdle model
specified by two distinct models (e.g. a binary and a truncated
count model) with the parameters of a non-hurdle count model.
No checks are made to ensure that the models supplied are comparable.
The function is designed for glmmADMB models only.
}
