% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cramers_v.R
\name{cramersv}
\alias{cramersv}
\title{Cramer's V}
\usage{
cramersv(x)
}
\arguments{
\item{x}{An abject created by \code{\link{chisq.test}}.}
}
\description{
This function calculates Cramer's V from an object created by the
\code{\link{chisq.test}} function from the \code{stats} package.
}
\examples{
cramersv(x)

cramersv(chisq.test(enve_test$extortion_victim, enve_test$size))
}
\keyword{association,}
\keyword{chi-square}
\keyword{frequency}
\keyword{measures}
\keyword{of}
\keyword{tables,}
\keyword{test}
