% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cramers_v.R
\name{cv_test}
\alias{cv_test}
\title{Cramer's V}
\usage{
cv_test(x)
}
\arguments{
\item{x}{An abject created by \code{\link{chisq.test}}.}
}
\description{
This function calculates Cramer's V from an abject created by the
\code{\link{chisq.test}} function from the \code{stats} package.
}
\examples{
cv_test(x)
}
\keyword{association,}
\keyword{chi-square}
\keyword{frequency}
\keyword{measures}
\keyword{of}
\keyword{tables,}
\keyword{test}
