% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion_batch.R
\name{dispersion_batch}
\alias{dispersion_batch}
\title{Dispersion index batch function}
\usage{
dispersion_batch(x, data, print_option = c("none", "markdown", "pandoc",
  "latex", "html"))
}
\arguments{
\item{x}{a vector of column names of counts in data}

\item{data}{a data frame}

\item{print_options}{Provides the option to print out results as
a table formatted by \code{\link{kable}} from the \code{knitr} package.}
}
\description{
A helper function to process batch index of dispersion tests and
add names of the variables to the results, default option to print
results as a matrix, can print as a kintr table
}
\examples{
dispersion_batch(x = c("extortions", "bribes"), data = testdata)
dispersion_batch(x = c("extortions", "bribes"), data = testdata,
                  print = "pandoc")
}
