% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion_index.R
\name{dispersion_index}
\alias{dispersion_index}
\title{Index of dispersion for count data}
\usage{
dispersion_index(x, data = NULL)
}
\arguments{
\item{x}{a vector of counts or the name of a column of counts in data}

\item{data}{a data frame}
}
\description{
Takes a vector and calculates the index of dispersion
Then evaluates its significance for Overdispersion only
Returns a list with named numbers
}
\examples{
dispersion_index("extortions", data = testdata)

lapply(c("extortions, "bribes), dispersion_index, data = testdata)
}
