% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fac_relevel.R
\name{fac_relevel}
\alias{fac_relevel}
\title{Auto releviling function for factors of count data}
\usage{
fac_relevel(x, min_levels = 5)
}
\arguments{
\item{x}{a numeric or integer vector.}

\item{max_levels}{offers the option to control the minimum
numer of levels in the new factor}
}
\description{
A helper function that turns a vector of count into factor data.
If there are more than 5 levels in the factor, it aggregates those
above the fifth level to a single level.
}
\examples{
fac_relevel(testdata$extortions)
}
\keyword{as.factor,}
\keyword{count}
\keyword{data}
\keyword{factor,}
