% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_glmmadmb.R
\name{get_glmmadmb}
\alias{get_glmmadmb}
\title{Get parameters from glmmADMB models}
\usage{
get_glmmadmb(mod)
}
\arguments{
\item{mod}{a model fitted with glmmADMB}
}
\description{
This function returns a list of model parameters from
models created with glmmADMB.
}
\examples{
require(glmmadmb)
model <- glmmadmb(extortions ~ 1, data = testdata, family = "poisson")

get_glmmadmb(model)
}
