% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurdle_compare.R
\name{hurdle_compare}
\alias{hurdle_compare}
\title{Compare parameters from hurdle models}
\usage{
hurdle_compare(..., modnames = NULL)
}
\arguments{
\item{...}{recieves the at least two \code{\link{joint_ic}}
functions with two models each (one binary and one count).}

\item{modnames}{option to add a character vector supplying a short
descriptive name for each pair of models}
}
\description{
Comapre parameters from at least two hurdle models supplied
as pairs in the \code{\link{joint_ic}} function.
}
