% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqv.R
\name{iqv}
\alias{iqv}
\title{Index of Qualitative Variation}
\usage{
iqv(x)
}
\arguments{
\item{x}{a dataframe with columns of counts for each k category}
}
\description{
A simple index of qualitative variation based on the equation in
\url{http://bit.ly/2qThZgd}. The equation is:
\deqn{IQV = \frac{k(100^2 - \sum p^2)}{100^2(k-1)}}
}
\examples{
years_by_state <- reshape2::dcast(testdata, CVE_ENT ~ yearsquant)
iqv(years_by_state)

# IQV per state
apply(years_by_state, 1, iqv)
}
