% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_ic.R
\name{joint_ic}
\alias{joint_ic}
\title{Joint hurdle parameters}
\usage{
joint_ic(m1, m2)
}
\arguments{
\item{m1}{A model}

\item{m2}{Another model}
}
\description{
Produces the joint paramters of a hurdle model based on
two distinct models: one modeling the binary part, and the
second modeling the truncated count part. Calculation of the
params is done simply by adding the paramters of the independent
models as suggested by Cameron and Trivedi (2013) and Hilbe (2011).
No checks are made to ensure models are indeed part of a hurdle equation.
}
