% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_test_batch.R
\name{ks_test_batch}
\alias{ks_test_batch}
\title{KS-Test batch}
\usage{
ks_test_batch(x, data, print_option = c("none", "markdown", "pandoc", "latex",
  "html"), ...)
}
\arguments{
\item{x}{a vector of column names of counts in data}

\item{data}{a data frame}

\item{print_options}{Provides the option to print out results as
a table formatted by \code{\link{kable}} from the \code{knitr} package.}
}
\description{
A helper function to process batch index of Kolmogorov-Smirnov Tests,
default option to print results as a matrix, can print as a kintr table
}
\examples{
ks_test_batch(c("extortions", "bribes") data = testdata,
                print_option = "pandoc")
}
