% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_gini_test.R
\name{mc_gini_test}
\alias{mc_gini_test}
\title{Monte Carlo Gini Index Test}
\usage{
mc_gini_test(x, data = NULL, reps = 2000, keep_reps = FALSE,
  family = c("poisson", "nbinom"), plots = FALSE)
}
\arguments{
\item{x}{A numeric or intenger vector of event counts. If
\code{data} is provided, x is the character label of the column
of event counts.}

\item{data}{A data frame.}

\item{reps}{The number of Monte Carlo replicates}

\item{keep_reps}{Logical option to keep the vector of Gini Index statistics
for the Monte Carlo replications.}

\item{family}{Family distribution for the Null Hypothesis.}

\item{plots}{Logical indicating whether to generate a plot of the test.}
}
\description{
Takes a vector of counts and calculates its Gini coefficient
then it generates a Monte Carlo distribution of Gini coefficients
under the specified family (Poisson or nbinom) and compares the
observed Gini with the MC confidence interval.
There is an option to generate a plot of the values and critical region.
}
\examples{
mc_gini_test("extortions", data = testdata, plots = TRUE, family = "poisson")
}
\keyword{Carlo,}
\keyword{Gini}
\keyword{Index}
\keyword{Monte}
\keyword{binomial,}
\keyword{inequality,}
\keyword{negative}
\keyword{poisson,}
\keyword{simulation,}
