% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_ks_test.R
\name{my_ks_test}
\alias{my_ks_test}
\title{My Kolmogorov-Smirnov Tests}
\usage{
my_ks_test(x, data = NULL, simulate.p.value = TRUE, B = 500,
  alternative = "greater", family = c("poisson", "nbinom"))
}
\arguments{
\item{x}{a vector of counts or the name of a column of counts in data}

\item{data}{a data frame}

\item{simulate.p.value}{logical to indicate if KS-test p-value should be
calculated using Monte Carlo simulation}

\item{B}{number of replicates of the null hypothesis distribution, and of
the KS-test p-value if simulate.p.value is \code{TRUE}}

\item{alternative}{Alternative hypothesis: The CDF is above the null}

\item{family}{the distribution of the null hypothesis}
}
\description{
Performs two-sample Kolmogorov-Smirnov tests based on the
\code{\link{dgof::ks.test}} from the \code{\link{dgof}} package
and the Monte Carlo replicates from the \code{\link{mc_gini_test}}
}
\examples{
my_ks_test(x = "extortions", data = testdata, family = "poisson")
}
