% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_kables.R
\name{print_kables}
\alias{print_kables}
\title{Print a list of kable tables}
\usage{
print_kables(x)
}
\arguments{
\item{x}{a list of kable tables}
}
\description{
Prints the kables direct to output, facilitating pickup by pandoc.
}
\examples{
b <- batch_chisq(df = testdata, DV = "extortion_victim",
            IV = c("bribe_victim", "size"))

l <- chisq_list(b, option = "ratio", print_option = "pandoc")

print_chisq_l(l)
}
