% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.zeroinfl.R
\name{update.zeroinfl}
\alias{update.zeroinfl}
\title{Update method for zeroinfl}
\usage{
\method{update}{zeroinfl}(object, new, ...)
}
\arguments{
\item{object}{zeroinfl model}

\item{new}{a formula representing the updated model}
}
\description{
A method for update function that works with zeroinfl models
from the pscl package. The code for this method was taken from
\url{https://stackoverflow.com/questions/21247078/updating-a-zeroinfl-model-in-r}.
}
\examples{
require(pscl)
model <- zeroinfl(extortions ~ bribes | bribes, dist = "negbin", data = testdata)

update(model, . ~ . | . + years)

}
