% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/victim_cumulative.R
\name{victim_cumulative}
\alias{victim_cumulative}
\title{Victimisation distribution cumulative percentages}
\usage{
victim_cumulative(data)
}
\arguments{
\item{x}{a data frame generated by \code{\link{victim_table}}}
}
\description{
A function to generate cumulative percentages from the
distribution provided by \code{\link{victim_table}}.
}
\examples{
vt <- victim_table("extortions_nas", data = testdata)
victim_cumulative(vt)

# Batch mode
vl <- lapply(c("extortions", "bribes"), victim_table, data = testdata)
lapply(vl, victim_cumulative)
}
