% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/victim_lorenz.R
\name{victim_lorenz}
\alias{victim_lorenz}
\title{Lorenz curves for victimisation distributions}
\usage{
victim_lorenz(x, data = NULL, family = c("none", "poisson", "nbinom"),
  reps = 500, by_var = NULL)
}
\arguments{
\item{x}{a vector of counts or the name of a column of counts in data}

\item{data}{a data frame}

\item{family}{an option based on \code{\link{gini_test}} that
adds curves for the expected counts under the null hypothesis
of a poisson or a negative binomial distribution using an aggregate
Monte Carlo simulation with 500 replicates}

\item{reps}{passes option of replicates to \code{\link{gini_test}}}
}
\description{
A wrapper for a series of functions to generate Lorenz plots from
vectors of counts of victimisation incidents. Includes option to include
Expected counts under Poisson or Negative Binomial expected counts.
}
\examples{
victim_lorenz("bribes", data = testdata, family = "poisson")

}
\keyword{Carlo,}
\keyword{Count}
\keyword{Curve,}
\keyword{Data}
\keyword{Lorenz}
\keyword{Monte}
