% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_data_deparser.R
\name{x_data_deparser}
\alias{x_data_deparser}
\title{x_data_deparser function}
\usage{
x_data_deparser(x, data = NULL)
}
\arguments{
\item{x}{a vector or a name of a column in \code{data}}

\item{data}{a data frame.}
}
\description{
A helper function used inside other functions to be able
to call either x as an object or as a column name of data
}
